Create a new recipe site with three recipes
A new recipe site is created in this commit for "The Odin Prjoect" to learn basic html. Main index page is created with links to other three recipe pages. Added images on each recipe page and back links to home page is created on each of them. Ingredients and steps for the recipes are added as content on each page.